Remove containment
I see why it was commented out in the first place.

The containment does not expand as the grid increases
in height. This prevents you from being able to stick the
item in the bottom row of the grid.

I looked into updating the containment dynamically with no success.
Better to not specify a contaiment for now.