Update README.md
Usually the installation section is one of the first you see when you check out a repo. Personally, I was about to download the source until I scrolled to the bottom. Just my two cents.. 

This is awesome, it's going to save me a bunch of time on my current project. Thanks for this!