Make just part of a widget draggable
If we add a parameter 'element' to 'draggable' config parameter, then, widget will be draggable just if it is under the html which has that class. 
This is useful for example if some widgets has header and we want to enable dragging of that widget only by its header. On that way we can avoid possible conflicts when trying to drag widget and clicking inside some controls in widget.

How does it work ? 
- if config draggable.element exists then, widget should have that class (value of draggable.element) somewhere in its html structure. User can than start dragging just from that html element in widget, or its child structure elements.