Fix drag scrolling to work with any container
When dragging an element offscreen, automatically scrolling of the
viewport only happens when the root-level document is the scroll
container.  If some other element within the page is the scroll
container, no automatic scrolling occurs, and it becomes impossible
to drag a gridster item to a position offscreen.

This fix replaces the current auto-scrolling mechanism with one which
identifies the nearest scrollable container to the element being
dragged and manipulates it's scroll viewport to create the automatic
scrolling effect (falling back to the document, if necessary).