Allow options.isMobile to dictate mobile state
The existing code ignores the `options.isMobile` flag and simply
computes what it thinks the value should be. This doesn't allow the
user of the library to decide upon their own way of detecting mobile
devices.

This change alters the `refresh` method to honor the `isMobile`
setting if present, and to compute it otherwise.