Don't allow items to be dragged down indefinitely
The current implementation allows items to be dragged well past the
bottom of the other elements. This causes the item to start to
accelerate and disappear as the underlying container also resizes
to contain the element.

This fix prevents the item from going past the bottom row of elements
to keep this from happening.