Fix drag bug in very short grids
An error in a prior commit made it so that an item dragged below the
bottom of the scroll container would make it become detached from the
cursor and speed off the bottom of the screen.

This change puts in a number of limits on how much the item can be
scrolled, and switches to using `getClientRects` for a more accurate
measure of the scroll container's bounds.